% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmer_analysis.R
\name{hmer_analysis}
\alias{hmer_analysis}
\title{create an hmer object}
\usage{
hmer_analysis(
  n_train = NULL,
  n_grps_in = NULL,
  groups_in = NULL,
  indx_out = NULL,
  full_groups_out = NULL
)
}
\description{
create an hmer object
}
